Nuevas mejoras a las vistas de productos banner-carruel, ligas playstation/switch y mejora a la imagen de usuario logueado
