+++ database/usersDataBase.json
@@ -1,38 +1,38 @@
 [
   {
     "id": 1,
-    "first_name": "Abraham",
-    "last_name": "Montelongo",
-    "email": "abraham@DH.com",
-    "password": "$2a$10$AjO827D2eMGb1ZpkhZ529u2S.IE05eJ8Tx3hXrJdxeIdcm6UbMdX6",
-    "category": "usuario",
-    "avatar": "1627059914027_img.jpg"
+    "first_name": "Cesar",
+    "last_name": "Villa DH",
+    "email": "cesar@DH.com",
+    "password": "$2a$10$L39e5Xu2CHcjGdAlOLwz6.8RMIIRFCDdspcgUlK72UR.sFmYBUiIS",
+    "category": "guest",
+    "avatar": "1628106840383_img.jpg"
   },
   {
     "id": 2,
-    "first_name": "Cesar",
-    "last_name": "Villa",
-    "email": "cesar@DH.com",
-    "password": "$2a$10$L39e5Xu2CHcjGdAlOLwz6.8RMIIRFCDdspcgUlK72UR.sFmYBUiIS",
-    "category": "invitado",
-    "avatar": "1627060004839_img.jpg"
-  },
-  {
-    "id": 3,
     "first_name": "Daniela",
-    "last_name": "Colin",
+    "last_name": "Mendoza Colin",
     "email": "dany@DH.com",
     "password": "$2a$10$SjkuVj8XjeF.h8SVM4.H8uLjzQJBdMyNdQQi9dHczqpIQ.bwieAgC",
-    "category": "usuario",
+    "category": "guest",
     "avatar": "1627071934044_img.jpg"
   },
   {
-    "id": 4,
+    "id": 3,
     "first_name": "lolita",
     "last_name": "lolita",
     "email": "lolita@DH.com",
     "password": "$2a$10$tu5Cljx2XY3jyBhq95mLP.3y0QGey8sOXxC282uAEywRKou.dGSJq",
     "category": "usuario",
     "avatar": "1627182191630_img.jpg"
+  },
+  {
+    "id": 4,
+    "first_name": "Abraham",
+    "last_name": "Montelongo",
+    "email": "abraham@DH.com",
+    "password": "$2a$10$jphKNznL.jvAYXAHo2u7me1WuWDfhp57rSWmA9gOuwiaCAhR2Y4VK",
+    "category": "usuario",
+    "avatar": "1628103044925_img.jpg"
   }
 ]
