+++ public/css/usersList.css
@@ -134,10 +134,17 @@ input[type="checkbox"]:checked + .navbar-hamburguer{
     background-color: #00db82;
     border-radius: 5px;
     padding: 5px;
+    height: 45px;
 }
 .fa-shopping-cart, .fa-user{
     color: white;
 }
+
+#imgUserLogged {
+    height: 100%;
+    border-radius: 100%;
+    border: solid white 1px;
+} 
 /* fin id login*/
 /*Fin Nuevo Header*/
