+++ views/index/index.ejs
@@ -7,17 +7,17 @@
 <body>
     <!--HEader-->
     <%- include('../partials/header');%>
+    <div id="userLogged">
+        <% if (locals.isLogged) { %>
+        <h1>Bienvenido: <%=userLogged.first_name%></h1>
+        <%}%>
+    </div>
     <!--navbar-desk-->
     <%- include('../partials/navbar-desk.ejs');%>
     <!--Banner-->
     <%- include('../partials/banner.ejs');%>    
     <!--Nuevo contenido main imagenes-->
     <main>
-        <div>
-            <% if (locals.isLogged) { %>
-            <h1 style="margin-left: 85px;">Hola: <%=userLogged.first_name%></h1>
-            <%}%>
-        </div>
         <div class="contendor-main">
             <%- include('../partials/playstation-product.ejs');%>
             <%- include('../partials/nintendoswitch-product.ejs');%>
