+++ views/partials/headerUser.ejs
@@ -25,11 +25,11 @@
         </form>                               
     </div>
     <form id="login">
-        <div class="contenedor-login-fas" id="user">
-            <a href="/users/login"><i class="far fa-user"></i></a>
-        </div>
         <div class="contenedor-login-fas" id="cart">
-            <a href="/users/list"><i class="fas fa-users"></i></a>
+            <a href="/users/<%=userLogged.id%>"><img id="imgUserLogged" src="/images/users/<%=userLogged.avatar%>" alt=""></a>
+        </div>
+        <div class="contenedor-login-fas"  id="user">
+            <a href="/users/logout"><i class="fas fa-sign-out-alt"></i></a>
         </div>
     </form>
 </header>
