+++ views/partials/navbar-desk.ejs
@@ -1,6 +1,6 @@
 <div id="content-navbar-desk">
     <ul id="navbar-desk">
-        <li><a href="/playstation">PlayStation</a></li>
-        <li><a href="/nintendo-switch">Nintendo Switch</a></li>
+        <li><a href="/playstation"><img src="/images/PlayStation.jpg" alt=""></a></li>
+        <li><a href="/nintendo-switch"><img src="/images/nintendo.jpg" alt=""></a></li>
     </ul>
 </div>
