+++ views/users/profile.ejs
@@ -32,10 +32,10 @@
                 </form>                               
             </div>
             <form id="login">
-                <div class="contenedor-login-fas" id="user">
-                    <a href="/users/<%=userLogged.id%>"><i class="far fa-user"></i></a>
-                </div>
                 <div class="contenedor-login-fas" id="cart">
+                    <a href="/users/<%=userLogged.id%>"><img id="imgUserLogged" src="/images/users/<%=userLogged.avatar%>" alt=""></a>
+                </div>
+                <div class="contenedor-login-fas"  id="user">
                     <a href="/users/logout"><i class="fas fa-sign-out-alt"></i></a>
                 </div>
             </form>
