+++ views/users/registerEdit.ejs
@@ -25,7 +25,25 @@
             </div>
             <!-- Con el enctype no puedo actualizar la informacion con el metodo PUT,
             pero si se lo quito el enctype se borra la imagen y no la pudo cargar de nuevo.-->
-            <form action="/<%= userToEdit.id %>?_method=PUT" method="POST" enctype="multipart/form-data">
+            <form action="/users/<%= userToEdit.id %>?_method=PUT" method="POST" enctype="multipart/form-data">
+                <div class="invoiceBox">
+                    <label for="">Foto de perfil</label>
+                    <div class="imagen">
+                        <figure><img  src="/images/users/<%=userToEdit.avatar%>" alt="imagen de usuario"></figure>
+                    </div>
+                    <label for="file">
+                      <div class="boxFile" >
+                        Cambiar foto
+                      </div>
+                    </label>
+                    <input  id="file" 
+                            name="avatar" 
+                            size="6000" 
+                            type="file" 
+                            value="<%= userToEdit.avatar %>"
+                            >
+                  </div>
+                
                 <label for="first_name">Nombre</label>
 			    <input  type="text" 
                         id="first_name" 
@@ -50,47 +68,7 @@
                         name="email" 
                         placeholder="Ej: evely@mail.com" 
                         class="form-input"
-                        value="<%= userToEdit.email%>"
-                        >
-
-                <label>Contraseña</label>
-                <div class="content-password">
-                    <input  type="password" 
-                            name="password" 
-                            id="password" 
-                            placeholder="Contraseña"
-                            value="<%= userToEdit.password %>"
-                            >
-                    <input id="check" type="checkbox" onclick="mostrarContrasena()">
-                    <span style="font-size: 10px; padding-left: 5px;">Mostrar</span>
-                </div>
-
-                <label for="category" class="form-label">Categoría:</label>
-				<select name="category" id="category" class="form-input">
-					<option value="">Elegir opción</option>
-                    <option value="user" selected="<%= userToEdit.category === 'user' ? 'selected' : '' %>">Usuario</option>
-                    <option value="guest" selected="<%= userToEdit.category === 'guest' ? 'selected' : '' %>">Invitado</option>
-				</select>
-
-                
-
-                <div class="invoiceBox">
-                    <label for="">Foto de perfil</label>
-                    <div class="imagen">
-                        <figure><img  src="/images/users/<%=userToEdit.avatar%>" alt="imagen de usuario"></figure>
-                    </div>
-                    <label for="file">
-                      <div class="boxFile" >
-                        Cambiar foto
-                      </div>
-                    </label>
-                    <input  id="file" 
-                            name="avatar" 
-                            size="6000" 
-                            type="file" 
-                            value="<%= userToEdit.avatar %>"
-                            >
-                  </div>
+                        value="<%= userToEdit.email%>">
 
                 <div class="btn-container">
                     <button class="btn-login" type="submit">GUARDAR CAMBIOS</button>
